Finish simple show of level in character panel
