--- ARPG-CSE5912-LTS/Assets/Scripts/LevelSystem/LevelShow.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1f407fd01e3dc9d40a12ee0bbea27928
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
