+++ ARPG-CSE5912-LTS/Assets/Scripts/LevelSystem/LvlExpShow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 038ecb40f7666b8409b338e96ea1552c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
